reload expired cache URLs
In some conditions, there can be a pause between the initial load of the
entry (that already contains a signed download URL even though we don't
need it yet) and the time the cache blob is actually downloaded at the end
of the build. One such case seems to be accessing SBOM scanner layers after
the main build is completed. The signed URL is only valid for 10 minutes,
and if more time has passed, then the old URL may receive a 403 error.

Signed-off-by: Tonis Tiigi <tonistiigi@gmail.com>